chore: refactors - package names, formatting
